abstract packet output cleanup
