+++ docs/testing/codestory-e2e-stats-log.md
@@ -81,6 +81,7 @@ Rows whose commit cell ends in `+wt` were run from the working tree based on tha
 | 2026-06-14 | 28717906+wt | pass, constrained packet claim profiles full-sidecar stats; proof_tier full_sidecar; warnings none; real drill intentionally skipped with CODESTORY_ALLOW_SKIP_REAL_REPO_DRILL_CASES=1; symbol_search_docs 12,420; dense anchors 725; dense skips 11,695; semantic_embedding_ms 45.88s; retrieval_index_seconds 6.47; retrieval_mode full; repeat full refresh 23.60s with 0 embedded; repeat graph 12.42s; repeat semantic 1.28s; repeat cache 4.97s; repeat search projection/index 0.96s/1.20s | 70.79 | 0.27 | 1.48 | 0.55 | 0.27 | 0.23 | 90,949 | 76,711 | 250 | 0 | 725 | true |
 | 2026-06-14 | 28717906+wt | pass, final constrained packet claim profiles with SWR acronym gate full-sidecar stats; proof_tier full_sidecar; warnings none; real drill intentionally skipped with CODESTORY_ALLOW_SKIP_REAL_REPO_DRILL_CASES=1; symbol_search_docs 12,420; dense anchors 725; dense skips 11,695; semantic_embedding_ms 45.60s; retrieval_index_seconds 6.47; retrieval_mode full; repeat full refresh 23.52s with 0 embedded; repeat graph 12.17s; repeat semantic 1.12s; repeat cache 4.86s; repeat search projection/index 0.97s/1.12s | 68.20 | 0.32 | 1.35 | 0.57 | 0.25 | 0.22 | 90,954 | 76,715 | 250 | 0 | 725 | true |
 | 2026-06-14 | 28717906+wt | pass, final cleanup without temporary plan docs full-sidecar stats; proof_tier full_sidecar; warnings none; real drill intentionally skipped with CODESTORY_ALLOW_SKIP_REAL_REPO_DRILL_CASES=1; symbol_search_docs 12,420; dense anchors 725; dense skips 11,695; semantic_embedding_ms 51.80s; retrieval_index_seconds 7.81; retrieval_mode full; repeat full refresh 24.41s with 0 embedded; repeat graph 12.66s; repeat semantic 0.68s; repeat cache 4.55s; repeat search projection/index 1.23s/1.11s | 76.13 | 0.31 | 1.48 | 0.71 | 0.28 | 0.23 | 90,954 | 76,715 | 250 | 0 | 725 | true |
+| 2026-06-14 | 69c033c4+wt | pass, packet output budget and trace-writer cleanup full-sidecar stats; proof_tier full_sidecar; warnings none; real drill intentionally skipped with CODESTORY_ALLOW_SKIP_REAL_REPO_DRILL_CASES=1; symbol_search_docs 12,425; dense anchors 725; dense skips 11,700; semantic_embedding_ms 46.17s; retrieval_index_seconds 6.90; retrieval_mode full; repeat full refresh 24.87s with 0 embedded; repeat graph 12.49s; repeat semantic 0.70s; repeat cache 5.92s; repeat search projection/index 1.02s/2.55s | 72.75 | 0.30 | 1.69 | 0.57 | 0.25 | 0.21 | 90,984 | 76,741 | 250 | 0 | 725 | true |
 
 ## Repeat And Report Timing
 
@@ -112,6 +113,7 @@ and zero-reembedding assertions are the actionable repeat-refresh gates.
 | 2026-06-14 | 28717906+wt | constrained packet claim profiles full-sidecar stats; proof_tier full_sidecar; real drill skipped with CODESTORY_ALLOW_SKIP_REAL_REPO_DRILL_CASES=1; repeat graph 12.42s; repeat semantic 1.28s; repeat cache/search projection/index 4.97s/0.96s/1.20s | 23.60 | 2.11 | 0.83 | 1.28 |
 | 2026-06-14 | 28717906+wt | final constrained packet claim profiles with SWR acronym gate full-sidecar stats; proof_tier full_sidecar; real drill skipped with CODESTORY_ALLOW_SKIP_REAL_REPO_DRILL_CASES=1; repeat graph 12.17s; repeat semantic 1.12s; repeat cache/search projection/index 4.86s/0.97s/1.12s | 23.52 | 2.07 | 0.82 | 1.24 |
 | 2026-06-14 | 28717906+wt | final cleanup without temporary plan docs full-sidecar stats; proof_tier full_sidecar; real drill skipped with CODESTORY_ALLOW_SKIP_REAL_REPO_DRILL_CASES=1; repeat graph 12.66s; repeat semantic 0.68s; repeat cache/search projection/index 4.55s/1.23s/1.11s | 24.41 | 2.19 | 0.92 | 1.27 |
+| 2026-06-14 | 69c033c4+wt | packet output budget and trace-writer cleanup full-sidecar stats; proof_tier full_sidecar; real drill skipped with CODESTORY_ALLOW_SKIP_REAL_REPO_DRILL_CASES=1; repeat graph 12.49s; repeat semantic 0.70s; repeat cache/search projection/index 5.92s/1.02s/2.55s | 24.87 | 2.03 | 0.80 | 1.23 |
 
 ## Phase Metrics
 
@@ -186,3 +188,4 @@ from this phase table rather than backfilled.
 | 2026-06-14 | 28717906+wt | constrained packet claim profiles full-sidecar stats; proof_tier full_sidecar; symbol_search_docs 12,420; dense anchors 725; dense skips 11,695; reasons public_api 669, entrypoint 6, central_graph_node 40, component_report 10 | 70.79 | 12.60 | 49.31 | 0 | 725 | 0 |
 | 2026-06-14 | 28717906+wt | final constrained packet claim profiles with SWR acronym gate full-sidecar stats; proof_tier full_sidecar; symbol_search_docs 12,420; dense anchors 725; dense skips 11,695; reasons public_api 669, entrypoint 6, central_graph_node 40, component_report 10 | 68.20 | 12.43 | 46.50 | 0 | 725 | 0 |
 | 2026-06-14 | 28717906+wt | final cleanup without temporary plan docs full-sidecar stats; proof_tier full_sidecar; symbol_search_docs 12,420; dense anchors 725; dense skips 11,695; reasons public_api 669, entrypoint 6, central_graph_node 40, component_report 10 | 76.13 | 12.43 | 53.24 | 0 | 725 | 0 |
+| 2026-06-14 | 69c033c4+wt | packet output budget and trace-writer cleanup full-sidecar stats; proof_tier full_sidecar; symbol_search_docs 12,425; dense anchors 725; dense skips 11,700; reasons public_api 669, entrypoint 6, central_graph_node 40, component_report 10 | 72.75 | 13.61 | 47.47 | 0 | 725 | 0 |
